+++ src/main/java/com/imsweb/naaccrxml/NaaccrXmlUtils.java
@@ -45,7 +45,7 @@ import com.imsweb.naaccrxml.runtime.NaaccrStreamConfiguration;
 public final class NaaccrXmlUtils {
 
     // specification version that this library implements (will be used when writing XML files)
-    public static final String CURRENT_SPECIFICATION_VERSION = SpecificationVersion.SPEC_1_6;
+    public static final String CURRENT_SPECIFICATION_VERSION = SpecificationVersion.SPEC_1_7;
 
     // structure tags in the XML
     public static final String NAACCR_XML_TAG_ROOT = "NaaccrData";