+++ src/test/java/com/imsweb/naaccrxml/sas/SasTest.java
@@ -51,8 +51,8 @@ public class SasTest {
             Assert.fail("Looks like a new format was added, please handle it in the SasXmlToCsv and SasCsvToXml constructors, and adjust this test: " + formats);
 
         //noinspection ConstantConditions
-        if (!"1.6".equals(NaaccrXmlUtils.CURRENT_SPECIFICATION_VERSION))
-            Assert.fail("Current implementation changed, please adjust it in the SasCsvToXml code, then change this test!");
+        if (!"1.7".equals(NaaccrXmlUtils.CURRENT_SPECIFICATION_VERSION))
+            Assert.fail("Current implementation changed, please adjust it in the SasFlatToXml and SasCsvToXml code, then change this test!");
     }
 
     @Test